% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoFlux.R
\name{PseudoM3Flux}
\alias{PseudoM3Flux}
\title{PseudoM3Flux}
\usage{
PseudoM3Flux(
  file = NULL,
  assay = c("RNA", "SCT"),
  values = c("pt", "st"),
  q_cutoff = 0.01,
  morans_cutoff = 0.05,
  cores = 4,
  pr_root = "Y_1",
  return_obj = T
)
}
\arguments{
\item{file}{Seurat object}

\item{assay}{Assays to choose}

\item{values}{Use pseudotime or spatialflux}

\item{q_cutoff}{q value cutoff}

\item{morans_cutoff}{morans values cutoff}

\item{cores}{CPU cores}

\item{pr_root}{Select random root}

\item{return_obj}{Return full object}
}
\description{
PseudoM3Flux
}
\details{
This function calculates pseudotime using Monocle3
}
