% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialFlux-package.R
\docType{package}
\name{SpatialFlux-package}
\alias{SpatialFlux-package}
\title{SpatialFlux: SpatialFlux: An R package for distance gradient analysis in spatial transcriptomics}
\description{
Spatial transcriptomics distance gradient analysis using Euclidean distances.
}
\keyword{internal}
