% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialFlux.R
\name{SpatialFluxHD}
\alias{SpatialFluxHD}
\title{SpatialFluxHD}
\usage{
SpatialFluxHD(file = NULL, reference = NULL, compare = NULL)
}
\arguments{
\item{file}{file containing clusters coordinates scaled by image factor}

\item{reference}{cluster as reference}

\item{compare}{clusters spots to compare against reference}
}
\description{
SpatialFluxHD
}
\details{
Calculates euclidean distance of Fiji reference lines and cluster establishing a distance gradient for Spatial HD
}
