% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial2Flux.R
\name{SpatialShinyFlux}
\alias{SpatialShinyFlux}
\title{SpatialShinyFlux}
\usage{
SpatialShinyFlux(file = NULL, file2 = NULL, id = NULL)
}
\arguments{
\item{file}{Seurat's cluster coordinates as input}

\item{file2}{Seurat coordinates of reference spots selected in ShinyApp}

\item{id}{Don't remember this}
}
\description{
SpatialShinyFlux
}
\details{
This function calculates euclidean distances of reference spots selected in ShinyApp against cluster coordinates of interest
}
